# backend/tests/data/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_redraw_sketch_api_notebook_entries__entry_id__sketch_put": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "scene": {
            "anyOf": [
              {
                "contentMediaType": "application/octet-stream",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scene"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_redraw_sketch_api_notebook_entries__entry_id__sketch_put",
        "type": "object"
      },
      "Body_upload_notebook_image_api_notebook__campaign_id__upload_post": {
        "properties": {
          "author_name": {
            "default": "Unknown",
            "title": "Author Name",
            "type": "string"
          },
          "author_type": {
            "default": "player",
            "title": "Author Type",
            "type": "string"
          },
          "character_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Character Id"
          },
          "content": {
            "default": "",
            "title": "Content",
            "type": "string"
          },
          "entry_type": {
            "default": "sketch",
            "title": "Entry Type",
            "type": "string"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "scene": {
            "anyOf": [
              {
                "contentMediaType": "application/octet-stream",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scene"
          },
          "title": {
            "default": "Attached Image",
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_notebook_image_api_notebook__campaign_id__upload_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NotebookEntryCreate": {
        "properties": {
          "author_name": {
            "title": "Author Name",
            "type": "string"
          },
          "author_type": {
            "title": "Author Type",
            "type": "string"
          },
          "character_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Character Id"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "entry_type": {
            "default": "field_log",
            "title": "Entry Type",
            "type": "string"
          },
          "image_data": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Data"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "visibility": {
            "default": "all",
            "title": "Visibility",
            "type": "string"
          }
        },
        "required": [
          "title",
          "content",
          "author_name",
          "author_type"
        ],
        "title": "NotebookEntryCreate",
        "type": "object"
      },
      "NotebookEntryResponse": {
        "properties": {
          "author_name": {
            "title": "Author Name",
            "type": "string"
          },
          "author_type": {
            "title": "Author Type",
            "type": "string"
          },
          "campaign_id": {
            "title": "Campaign Id",
            "type": "integer"
          },
          "character_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Character Id"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "entry_type": {
            "default": "field_log",
            "title": "Entry Type",
            "type": "string"
          },
          "has_scene": {
            "default": false,
            "title": "Has Scene",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "image_data": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Data"
          },
          "ink_color": {
            "title": "Ink Color",
            "type": "string"
          },
          "is_deleted": {
            "default": false,
            "title": "Is Deleted",
            "type": "boolean"
          },
          "page_number": {
            "title": "Page Number",
            "type": "integer"
          },
          "pen_font": {
            "title": "Pen Font",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "visibility": {
            "default": "all",
            "title": "Visibility",
            "type": "string"
          }
        },
        "required": [
          "id",
          "campaign_id",
          "character_id",
          "author_name",
          "author_type",
          "pen_font",
          "ink_color",
          "title",
          "content",
          "created_at",
          "page_number"
        ],
        "title": "NotebookEntryResponse",
        "type": "object"
      },
      "NotebookEntryUpdate": {
        "properties": {
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "NotebookEntryUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/notebook/entries/{entry_id}": {
      "delete": {
        "operationId": "delete_notebook_entry_api_notebook_entries__entry_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Notebook Entry"
      },
      "put": {
        "operationId": "update_notebook_entry_api_notebook_entries__entry_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotebookEntryUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotebookEntryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Notebook Entry"
      }
    },
    "/api/notebook/entries/{entry_id}/scene": {
      "get": {
        "operationId": "get_sketch_scene_api_notebook_entries__entry_id__scene_get",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Sketch Scene"
      }
    },
    "/api/notebook/entries/{entry_id}/sketch": {
      "put": {
        "operationId": "redraw_sketch_api_notebook_entries__entry_id__sketch_put",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_redraw_sketch_api_notebook_entries__entry_id__sketch_put"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotebookEntryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Redraw Sketch"
      }
    },
...
    "/api/notebook/{campaign_id}/entries": {
      "get": {
        "operationId": "fetch_notebook_entries_api_notebook__campaign_id__entries_get",
        "parameters": [
          {
            "in": "path",
            "name": "campaign_id",
            "required": true,
            "schema": {
              "title": "Campaign Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "role",
            "required": false,
            "schema": {
              "default": "player",
              "title": "Role",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "character_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Character Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NotebookEntryResponse"
                  },
                  "title": "Response Fetch Notebook Entries Api Notebook  Campaign Id  Entries Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fetch Notebook Entries"
      },
      "post": {
        "operationId": "add_notebook_entry_api_notebook__campaign_id__entries_post",
        "parameters": [
          {
            "in": "path",
            "name": "campaign_id",
            "required": true,
            "schema": {
              "title": "Campaign Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotebookEntryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotebookEntryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Notebook Entry"
      }
    },
    "/api/notebook/{campaign_id}/upload": {
      "post": {
        "operationId": "upload_notebook_image_api_notebook__campaign_id__upload_post",
        "parameters": [
          {
            "in": "path",
            "name": "campaign_id",
            "required": true,
            "schema": {
              "title": "Campaign Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_notebook_image_api_notebook__campaign_id__upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Notebook Image"
      }
    },
...
  }
}